ai_assistant/requirements.txt
# Основные зависимости
requests>=2.31.0
Pillow>=10.0.0

# Для работы с UI
pyautogui>=0.9.54
pygetwindow>=0.0.9

# Для горячих клавиш (глобальный перехват)
pynput>=1.7.6

# Для Windows-специфичных функций (опционально)
pywin32>=306
winshell>=0.6

# Для работы с Excel, Word, PowerPoint через COM
# pywin32 уже включает необходимые компоненты
